Remove toString when decoding dictionary pages
This string conversion was causing issues with decoding
FIXED_LEN_BYTE_ARRAY and other Buffer based values due
because the default encoding used is utf-8 which is lossy.

This change should make decoding dictionary pages behave
closer to other page types, and will result in values
being returned closer to the types defined in the schema.

The corresponding changes in tests are all places where the
a dictionary page was being decoded and now the values are
being returned as either the numeric value, or buffer value
that they represented.